[FIX] unbind url
