+++ shopinvader/models/shopinvader_product.py
@@ -42,6 +42,11 @@ class ShopinvaderProduct(models.Model):
     shopinvader_display_name = fields.Char(
         compute='_compute_name', readonly=True
     )
+    shopinvader_categ_ids = fields.Many2many(
+        comodel_name='shopinvader.category',
+        compute='_compute_shopinvader_category',
+        string='Shopinvader Categories')
+
 
     _sql_constraints = [
         ('record_uniq', 'unique(backend_id, record_id, lang_id)',
@@ -61,6 +66,24 @@ class ShopinvaderProduct(models.Model):
         self.filtered(
             lambda p: not p.active).mapped('shopinvader_variant_ids').write(
                 {'active': False})
+
+    def _get_categories(self):
+        self.ensure_one()
+        return self.categ_id
+
+    def _compute_shopinvader_category(self):
+        for record in self:
+            ids = []
+            categs = record._get_categories()
+            for categ in categs:
+                parents = self.env['shopinvader.category'].search([
+                    ('parent_left', '<=', categ.parent_left),
+                    ('parent_right', '>=', categ.parent_right),
+                    ('backend_id', '=', record.backend_id.id),
+                    ('lang_id', '=', record.lang_id.id),
+                    ])
+                ids += parents.ids
+            record.shopinvader_categ_ids = ids
 
     def _prepare_shopinvader_variant(self, variant):
         values = {
@@ -136,3 +159,21 @@ class ShopinvaderProduct(models.Model):
                 'active': False,
             })
         return {}
+
+    def _redirect_existing_url(self):
+        """
+        During unbind, we have to redirect existing urls to the (first) related
+        shopinvader category.
+        :return: bool
+        """
+        for record in self.filtered(lambda p: p.url_url_ids):
+            # Active category without children
+            categs = record.shopinvader_categ_ids.filtered(
+                lambda c: c.active and not c.shopinvader_child_ids)
+            if categs:
+                categ = fields.first(categs)
+                record.url_url_ids.write({
+                    'redirect': True,
+                    'model_id': "%s,%s" % (categ._name, categ.id),
+                })
+        return True
