+++ shopinvader_algolia/models/__init__.py
@@ -1,3 +1,3 @@
 # -*- coding: utf-8 -*-
 
-from . import shopinvader_variant
+from . import shopinvader_product
