+++ shopinvader_locomotive/models/__init__.py
@@ -3,4 +3,3 @@ from . import shopinvader_backend
 from . import locomotive_binding
 from . import shopinvader_partner
 from . import keychain
-from . import shopinvader_variant
