+++ shopinvader_search_engine/models/shopinvader_backend.py
@@ -29,3 +29,8 @@ class ShopinvaderBackend(models.Model):
         for index in self.mapped('se_backend_id.index_ids'):
             index.clear_index()
         return True
+
+    @api.multi
+    def force_resynchronize_index(self):
+        self.mapped('se_backend_id.index_ids').resynchronize_all_bindings()
+        return True
